data_simulator/requirements.txt
pyarrow>=8.0.0
pandas>=1.3.0
ruamel.yaml>=0.17.10
aiokafka>=0.8.0
fastparquet>=0.7.1